+++ beep/tests/test_dataset.py
@@ -11,19 +11,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Unit tests related to feature generation"""
+"""Unit tests related to dataset generation"""
 
 import unittest
 import os
-import json
-import numpy as np
 from beep.featurize import (
     RPTdQdVFeatures,
     HPPCResistanceVoltageFeatures,
     DiagnosticSummaryStats,
+    DiagnosticProperties
 )
-from beep.dataset import BeepDataset
+from beep import MODULE_DIR
+from beep.dataset import BeepDataset, get_threshold_targets
 from monty.tempfile import ScratchDir
+from monty.serialization import dumpfn, loadfn
+import shutil
 
 TEST_DIR = os.path.dirname(__file__)
 TEST_FILE_DIR = os.path.join(TEST_DIR, "test_files")
@@ -33,7 +35,9 @@ FASTCHARGE_PROCESSED = os.path.join(TEST_FILE_DIR, '2017-06-30_2C-10per_6C_CH10_
 BIG_FILE_TESTS = os.environ.get("BEEP_BIG_TESTS", False)
 SKIP_MSG = "Tests requiring large files with diagnostic cycles are disabled, set BIG_FILE_TESTS to run full tests"
 FEATURIZER_CLASSES = [RPTdQdVFeatures, HPPCResistanceVoltageFeatures, DiagnosticSummaryStats]
-
+FEATURE_HYPERPARAMS = loadfn(
+    os.path.join(MODULE_DIR, "features/feature_hyperparameters.yaml")
+)
 
 class TestDataset(unittest.TestCase):
     def setUp(self):
@@ -50,9 +54,37 @@ class TestDataset(unittest.TestCase):
         self.assertSetEqual(set(dataset.feature_sets.keys()), {'RPTdQdVFeatures', 'DiagnosticSummaryStats'})
         self.assertEqual(dataset.missing.feature_class.iloc[0], 'HPPCResistanceVoltageFeatures')
 
+    def test_serialization(self):
+        with ScratchDir("."):
+            os.environ["BEEP_PROCESSING_DIR"] = os.getcwd()
+            dataset = BeepDataset.from_features('test_dataset', ['PreDiag'], FEATURIZER_CLASSES,
+                                            feature_dir=os.path.join(TEST_FILE_DIR, 'data-share/features'))
+            dumpfn(dataset, 'temp_dataset.json')
+            dataset = loadfn('temp_dataset.json')
+            self.assertEqual(dataset.name, 'test_dataset')
+            self.assertEqual(dataset.data.shape, (2, 56))
+            # from pdb import set_trace; set_trace()
+            self.assertListEqual(list(dataset.data.seq_num), [196, 197])
+            self.assertIsNone(dataset.X_test)
+            self.assertSetEqual(set(dataset.feature_sets.keys()), {'RPTdQdVFeatures', 'DiagnosticSummaryStats'})
+            self.assertEqual(dataset.missing.feature_class.iloc[0], 'HPPCResistanceVoltageFeatures')
+            self.assertIsInstance(dataset.filenames, list)
+
+            os.environ["BEEP_PROCESSING_DIR"] = os.getcwd()
+            dataset2 = BeepDataset.from_features('test_dataset', ['PreDiag'], [RPTdQdVFeatures],
+                                                feature_dir=os.path.join(TEST_FILE_DIR, 'data-share/features'))
+            dumpfn(dataset2, "temp_dataset_2.json")
+            dataset2 = loadfn('temp_dataset_2.json')
+            self.assertEqual(dataset2.missing.columns.to_list(), ["filename", "feature_class"])
+
     def test_from_processed_cycler_run_list(self):
         with ScratchDir("."):
             os.environ["BEEP_PROCESSING_DIR"] = os.getcwd()
+            os.makedirs(os.path.join(os.getcwd(), "data-share", "raw", "parameters"))
+            parameter_files = os.listdir(os.path.join(TEST_FILE_DIR, "data-share", "raw", "parameters"))
+            for file in parameter_files:
+                shutil.copy(os.path.join(TEST_FILE_DIR, "data-share", "raw", "parameters", file),
+                            os.path.join(os.getcwd(), "data-share", "raw", "parameters"))
             dataset = BeepDataset.from_processed_cycler_runs('test_dataset',
                                                              project_list=None,
                                                              processed_run_list=[DIAGNOSTIC_PROCESSED,
@@ -66,6 +98,39 @@ class TestDataset(unittest.TestCase):
             self.assertIsNone(dataset.X_test)
 
             self.assertEqual(dataset.missing.shape, (3, 2))
+            self.assertEqual(dataset.missing.filename.iloc[0],
+                             os.path.split(FASTCHARGE_PROCESSED)[1])
+
+    def test_dataset_with_custom_feature_hyperparameters(self):
+        with ScratchDir("."):
+            os.environ["BEEP_PROCESSING_DIR"] = os.getcwd()
+            os.makedirs(os.path.join(os.getcwd(), "data-share", "raw", "parameters"))
+            parameter_files = os.listdir(os.path.join(TEST_FILE_DIR, "data-share", "raw", "parameters"))
+            for file in parameter_files:
+                shutil.copy(os.path.join(TEST_FILE_DIR, "data-share", "raw", "parameters", file),
+                            os.path.join(os.getcwd(), "data-share", "raw", "parameters"))
+            hyperparameter_dict = {'RPTdQdVFeatures': [
+                {'diag_ref': 0, 'diag_nr': 1, 'charge_y_n': 0, 'rpt_type': 'rpt_0.2C', 'plotting_y_n': 0},
+                {'diag_ref': 0, 'diag_nr': 1, 'charge_y_n': 0, 'rpt_type': 'rpt_1C', 'plotting_y_n': 0},
+                {'diag_ref': 0, 'diag_nr': 1, 'charge_y_n': 0, 'rpt_type': 'rpt_2C', 'plotting_y_n': 0}],
+                                   'HPPCResistanceVoltageFeatures': [
+                                       FEATURE_HYPERPARAMS['HPPCResistanceVoltageFeatures']],
+                                   'DiagnosticSummaryStats': [FEATURE_HYPERPARAMS['DiagnosticSummaryStats']]
+                                   }
+            dataset = BeepDataset.from_processed_cycler_runs('test_dataset',
+                                                             project_list=None,
+                                                             processed_run_list=[DIAGNOSTIC_PROCESSED,
+                                                                                 FASTCHARGE_PROCESSED],
+                                                             feature_class_list=FEATURIZER_CLASSES,
+                                                             processed_dir=TEST_FILE_DIR,
+                                                             hyperparameter_dict=hyperparameter_dict,
+                                                             feature_dir='data-share/features')
+            self.assertEqual(dataset.name, 'test_dataset')
+            self.assertEqual(dataset.data.shape, (1, 135))
+            self.assertEqual(dataset.data.seq_num.iloc[0], 240)
+            self.assertIsNone(dataset.X_test)
+
+            self.assertEqual(dataset.missing.shape, (5, 2))
             self.assertEqual(dataset.missing.filename.iloc[0],
                              os.path.split(FASTCHARGE_PROCESSED)[1])
 
@@ -107,3 +172,32 @@ class TestDataset(unittest.TestCase):
                           }
 
         self.assertDictEqual(dataset.train_cells_parameter_dict, parameter_dict)
+
+    def test_get_threshold_targets(self):
+        dataset_diagnostic_properties = loadfn(os.path.join(TEST_FILE_DIR, "diagnostic_properties_test.json"))
+        threshold_targets_df = get_threshold_targets(dataset_diagnostic_properties.data,
+                                                     cycle_type="rpt_1C")
+        self.assertEqual(len(threshold_targets_df), 92)
+        self.assertEqual(threshold_targets_df.columns.to_list(), ['file',
+                                                                  'seq_num',
+                                                                  'initial_regular_throughput',
+                                                                  'rpt_1Cdischarge_energy0.8_normalized_reg_throughput',
+                                                                  'rpt_1Cdischarge_energy0.8_real_reg_throughput',
+                                                                  'rpt_1Cdischarge_energy0.8_cycles']
+                         )
+        self.assertEqual(threshold_targets_df[threshold_targets_df['seq_num'] == 154].round(decimals=3).to_dict("list"),
+                         {
+                             'file': ['PredictionDiagnostics_000154'],
+                             'seq_num': [154],
+                             'initial_regular_throughput': [489.31],
+                             'rpt_1Cdischarge_energy0.8_normalized_reg_throughput': [4.453],
+                             'rpt_1Cdischarge_energy0.8_real_reg_throughput': [2178.925],
+                             'rpt_1Cdischarge_energy0.8_cycles': [159.766]
+                          }
+                         )
+        threshold_targets_df = get_threshold_targets(dataset_diagnostic_properties.data,
+                                                     cycle_type="rpt_1C",
+                                                     extrapolate_threshold=False)
+        self.assertEqual(len(threshold_targets_df), 64)
+        self.assertEqual(threshold_targets_df['rpt_1Cdischarge_energy0.8_real_reg_throughput'].round(decimals=3)
+                         .median(), 2016.976)