+++ beep/tests/test_features.py
@@ -30,9 +30,12 @@ from beep.featurize import (
     DiagnosticProperties,
     DiagnosticSummaryStats,
 )
+from beep.structure import RawCyclerRun
 from beep.features import featurizer_helpers
+from beep.utils import parameters_lookup
 from monty.serialization import dumpfn, loadfn
 from monty.tempfile import ScratchDir
+from beep.utils.s3 import download_s3_object
 
 TEST_DIR = os.path.dirname(__file__)
 TEST_FILE_DIR = os.path.join(TEST_DIR, "test_files")
@@ -43,7 +46,7 @@ MACCOR_FILE_W_PARAMETERS = os.path.join(
     TEST_FILE_DIR, "PredictionDiagnostics_000109_tztest.010"
 )
 
-BIG_FILE_TESTS = os.environ.get("BEEP_BIG_TESTS", False)
+BIG_FILE_TESTS = os.environ.get("BIG_FILE_TESTS", None) == "True"
 SKIP_MSG = "Tests requiring large files with diagnostic cycles are disabled, set BIG_FILE_TESTS to run full tests"
 
 
@@ -306,6 +309,121 @@ class TestFeaturizer(unittest.TestCase):
                 ["ohmic_r_d0", "D_8"],
             )
 
+    def test_get_hppc_ocv(self):
+        pcycler_run_loc = os.path.join(
+            TEST_FILE_DIR, "PreDiag_000240_000227_truncated_structure.json"
+        )
+        os.environ["BEEP_PROCESSING_DIR"] = TEST_FILE_DIR
+        pcycler_run = loadfn(pcycler_run_loc)
+        hppc_ocv_features = featurizer_helpers.get_hppc_ocv(pcycler_run, 1)
+        self.assertEqual(np.round(hppc_ocv_features['variance of ocv'].iloc[0], 6), 0.000016)
+
+    def test_get_step_index(self):
+        pcycler_run_loc = os.path.join(
+            TEST_FILE_DIR, "PreDiag_000240_000227_truncated_structure.json"
+        )
+
+        parameters_path = os.path.join(TEST_FILE_DIR, "data-share", "raw", "parameters")
+        os.environ["BEEP_PROCESSING_DIR"] = TEST_FILE_DIR
+        pcycler_run = loadfn(pcycler_run_loc)
+        data = pcycler_run.diagnostic_interpolated
+        hppc_cycles = data.loc[data.cycle_type == "hppc"]
+        print(hppc_cycles.step_index.unique())
+        _, protocol_name = os.path.split(pcycler_run.protocol)
+        parameter_row, _ = parameters_lookup.get_protocol_parameters(protocol_name, parameters_path=parameters_path)
+
+        for cycle in hppc_cycles.cycle_index.unique():
+            hppc_cycle = hppc_cycles[hppc_cycles.cycle_index == cycle]
+            for step in hppc_cycle.step_index.unique():
+                hppc_cycle_step = hppc_cycle[(hppc_cycle.step_index == step)]
+                for step_iter in hppc_cycle_step.step_index_counter.unique():
+                    hppc_cycle_step_iter = hppc_cycle_step[(hppc_cycle_step.step_index_counter == step_iter)]
+                    duration = hppc_cycle_step_iter.test_time.max() - hppc_cycle_step_iter.test_time.min()
+                    median_crate = np.round(hppc_cycle_step.current.median() /
+                                            parameter_row["capacity_nominal"].iloc[0], 2)
+                    print(step, median_crate, duration)
+
+        step_ind = featurizer_helpers.get_step_index(pcycler_run,
+                                                     cycle_type="hppc",
+                                                     diag_pos=0)
+        self.assertEqual(len(step_ind.values()), 6)
+        print([step_ind["hppc_long_rest"],
+               step_ind["hppc_discharge_pulse"],
+               step_ind["hppc_short_rest"],
+               step_ind["hppc_charge_pulse"],
+               step_ind["hppc_discharge_to_next_soc"]])
+
+        self.assertEqual(step_ind, {
+            'hppc_charge_to_soc': 9,
+            'hppc_long_rest': 11,
+            'hppc_discharge_pulse': 12,
+            'hppc_short_rest': 13,
+            'hppc_charge_pulse': 14,
+            'hppc_discharge_to_next_soc': 15
+        })
+        step_ind = featurizer_helpers.get_step_index(pcycler_run,
+                                                     cycle_type="hppc",
+                                                     diag_pos=1)
+        self.assertEqual(len(step_ind.values()), 6)
+        self.assertEqual(step_ind, {
+            'hppc_charge_to_soc': 41,
+            'hppc_long_rest': 43,
+            'hppc_discharge_pulse': 44,
+            'hppc_short_rest': 45,
+            'hppc_charge_pulse': 46,
+            'hppc_discharge_to_next_soc': 47
+        })
+
+    def test_get_step_index_2(self):
+        pcycler_run_loc = os.path.join(
+            TEST_FILE_DIR, "PreDiag_000400_000084_truncated_structure.json"
+        )
+        parameters_path = os.path.join(TEST_FILE_DIR, "data-share", "raw", "parameters")
+        os.environ["BEEP_PROCESSING_DIR"] = TEST_FILE_DIR
+        pcycler_run = loadfn(pcycler_run_loc)
+        _, protocol_name = os.path.split(pcycler_run.protocol)
+        parameter_row, _ = parameters_lookup.get_protocol_parameters(protocol_name, parameters_path=parameters_path)
+
+        step_ind = featurizer_helpers.get_step_index(pcycler_run,
+                                                     cycle_type="hppc",
+                                                     diag_pos=0)
+        self.assertEqual(len(step_ind.values()), 7)
+
+        self.assertEqual(step_ind, {
+            'hppc_charge_to_soc': 9,
+            'hppc_long_rest': 11,
+            'hppc_discharge_pulse': 12,
+            'hppc_short_rest': 13,
+            'hppc_charge_pulse': 14,
+            'hppc_discharge_to_next_soc': 15,
+            'hppc_final_discharge': 17
+        })
+        step_ind = featurizer_helpers.get_step_index(pcycler_run,
+                                                     cycle_type="hppc",
+                                                     diag_pos=1)
+        self.assertEqual(len(step_ind.values()), 7)
+        self.assertEqual(step_ind, {
+            'hppc_charge_to_soc': 41,
+            'hppc_long_rest': 43,
+            'hppc_discharge_pulse': 44,
+            'hppc_short_rest': 45,
+            'hppc_charge_pulse': 46,
+            'hppc_discharge_to_next_soc': 47,
+            'hppc_final_discharge': 49
+        })
+
+    def test_get_diffusion_coeff(self):
+        with ScratchDir("."):
+            os.environ["BEEP_PROCESSING_DIR"] = TEST_FILE_DIR
+            pcycler_run_loc = os.path.join(
+                TEST_FILE_DIR, "PreDiag_000240_000227_truncated_structure.json"
+            )
+            pcycler_run = loadfn(pcycler_run_loc)
+            diffusion_df = featurizer_helpers.get_diffusion_coeff(pcycler_run, 1)
+            print(np.round(diffusion_df.iloc[0].to_list(), 3))
+            self.assertEqual(np.round(diffusion_df.iloc[0].to_list(), 3)[0], -0.016)
+            self.assertEqual(np.round(diffusion_df.iloc[0].to_list(), 3)[5], -0.011)
+
     def test_HPPCRelaxationFeatures_class(self):
         with ScratchDir("."):
             os.environ["BEEP_PROCESSING_DIR"] = TEST_FILE_DIR
@@ -353,6 +471,22 @@ class TestFeaturizer(unittest.TestCase):
                 [featurizer.X.columns[0], featurizer.X.columns[-1]],
                 ["var_charging_capacity", "square_discharging_dQdV"],
             )
+            x = [-3.622991274215596, -1.4948801528128568, -2.441732890889216, -0.794422489658189, 0.4889470327970021,
+                 0.7562360890191123, -0.9122534588595697, -3.771727344982484, -1.6613278517299095, -3.9279757071656616,
+                 0.1418911233780052, 0.7493913209640308, 0.6755655006191633, -1.0823827139302122, -2.484906394983077,
+                 -0.8949449222504844, -1.7523322777749897, -1.4575307327423712, 0.4889470327970021, 1.3265006178265961,
+                 0.2422557417274141, -2.6373799375134594, -1.230847957965504, -2.046540216421213, 0.2334339752067063,
+                 0.8239822694093881, 1.2085578295115413, 0.06687710057927358, -1.0135736732168983, 0.12101479889802537,
+                 -2.2735196264247866, 0.37844357940755063, 0.4889470327970021, 1.8786507359201035, 1.6731897281287798,
+                 -1.1875358619917917, 0.1361208058450041, -1.8275104616090456, -0.2665523054105704, 1.1375831683815445,
+                 1.84972885518774, 1.5023615714170622]
+            computed = featurizer.X.iloc[0].tolist()
+            for indx, value in enumerate(x):
+                precision = 6
+                self.assertEqual(np.round(value, precision), np.round(computed[indx], precision))
+
+            self.assertEqual(np.round(featurizer.X['var_discharging_capacity'].iloc[0], 3),
+                             np.round(-3.771727344982484, 3))
 
     def test_DiagnosticProperties_class(self):
         with ScratchDir("."):
@@ -479,3 +613,48 @@ class TestFeaturizer(unittest.TestCase):
             print(v_vars_df)
             self.assertEqual(np.round(v_vars_df.iloc[0]['var(v_diff)'], decimals=8),
                              np.round(9.71e-06, decimals=8))
+
+
+class TestRawToFeatures(unittest.TestCase):
+    def setUp(self):
+        self.maccor_file_w_parameters_s3 = {
+            "bucket": "beep-sync-test-stage",
+            "key": "big_file_tests/PreDiag_000287_000128.092"
+        }
+        self.maccor_file_w_parameters = os.path.join(
+            TEST_FILE_DIR, "PreDiag_000287_000128.092"
+        )
+
+    @unittest.skipUnless(BIG_FILE_TESTS, SKIP_MSG)
+    def test_raw_to_features(self):
+        os.environ["BEEP_PROCESSING_DIR"] = TEST_FILE_DIR
+
+        download_s3_object(bucket=self.maccor_file_w_parameters_s3["bucket"],
+                           key=self.maccor_file_w_parameters_s3["key"],
+                           destination_path=self.maccor_file_w_parameters)
+
+        with ScratchDir("."):
+            os.environ["BEEP_PROCESSING_DIR"] = TEST_FILE_DIR
+            # os.environ['BEEP_PROCESSING_DIR'] = os.getcwd()
+            cycler_run = RawCyclerRun.from_file(self.maccor_file_w_parameters)
+            processed_cycler_run = cycler_run.to_processed_cycler_run()
+            processed_cycler_run_path = os.path.join(
+                TEST_FILE_DIR, "processed_diagnostic.json"
+            )
+            # Dump to the structured file and check the file size
+            dumpfn(processed_cycler_run, processed_cycler_run_path)
+            # Create dummy json obj
+            json_obj = {
+                "file_list": [processed_cycler_run_path],
+                "run_list": [0],
+            }
+            json_string = json.dumps(json_obj)
+
+            newjsonpaths = process_file_list_from_json(
+                json_string, processed_dir=os.getcwd()
+            )
+            reloaded = json.loads(newjsonpaths)
+            result_list = ['success', 'success', 'success', 'success', 'success', 'success', 'success']
+            self.assertEqual(reloaded['result_list'], result_list)
+            rpt_df = loadfn(reloaded['file_list'][0])
+            self.assertEqual(np.round(rpt_df.X['m0_Amp_rpt_0.2C_1'].iloc[0], 6), 0.867371)